Improve metadata, and start grapher step for supply
