--- etl/steps/data/grapher/iea/2024-07-04/critical_minerals.py
@@ -1,32 +1,0 @@
-"""Load a garden dataset and create a grapher dataset."""
-
-from etl.helpers import PathFinder, create_dataset
-
-# Get paths and naming conventions for current step.
-paths = PathFinder(__file__)
-
-
-def run(dest_dir: str) -> None:
-    #
-    # Load inputs.
-    #
-    # Load garden dataset.
-    ds_garden = paths.load_dataset("critical_minerals")
-
-    # Read table from garden dataset.
-    tb = ds_garden["critical_minerals"]
-
-    #
-    # Process data.
-    #
-
-    #
-    # Save outputs.
-    #
-    # Create a new grapher dataset with the same metadata as the garden dataset.
-    ds_grapher = create_dataset(
-        dest_dir, tables=[tb], check_variables_metadata=True, default_metadata=ds_garden.metadata
-    )
-
-    # Save changes in the new grapher dataset.
-    ds_grapher.save()
